// lesson-13/src/models/inventory.dto.ts
export interface InventoryDto {
    '8': number;
    sold: number;
    string: number;
    'sold ': number;
    d: number;
    unavailable: number;
    pending: number;
    available: number;
    weisskeiner1: number;
    '¸E\u0011􉁶¦': number;
    availabl: number;
    peric: number;
    Sold: number;
    blocked: number;
    '{{status}}': number;
    available_DEV: number;
    lost: number;
}